:fire: custom element for cursor highlights
We didn't need a custom element for that. Worse, the current
implementation was holding a reference to the marker/decoration/editor
forever as these properties were never released (no delete, nothing).
This can be problematic as Atom won't be able to free that memory
properly.

Also, it was way simpler to just create a new decoration for error
rather than updating the previous one. Again, keeping references around
for this is suboptimal. I don't think it makes a big difference
visually, but maybe there's some adjustments to make.
