+++ lib/elements/navigable-panel.js
@@ -162,10 +162,7 @@ class NavigablePanel extends HTMLElement {
       // visible, we turn it red
       atom.beep();
 
-      let expandAtCursorHighlight = Kite.getLastExpandAtCursorHighlight();
-      if (expandAtCursorHighlight) {
-        expandAtCursorHighlight.indicateError();
-      }
+      Kite.highlightWordAtPosition(editor, position, 'error');
     });
   }
