+++ lib/kite.js
@@ -10,8 +10,7 @@ let child_process, CompositeDisposable, TextEditor, AccountManager,
   KiteStatus, KiteStatusPanel, KiteSidebar, KiteErrorRescueSidebar, NotificationsCenter,
   RollbarReporter, OverlayManager, KiteEditor, DataLoader,
   DisposableEvent, KiteSearch, EditorEvents, BrowserClient,
-  addDelegatedEventListener, ERROR_RESCUE_SHOW_SIDEBAR,
-  KiteExpandAtCursorHighlight;
+  addDelegatedEventListener, ERROR_RESCUE_SHOW_SIDEBAR, VirtualCursor;
 
 const Kite = module.exports = {
   activate() {
@@ -29,7 +28,6 @@ const Kite = module.exports = {
       NotificationsCenter = require('./notifications-center');
       RollbarReporter = require('./rollbar-reporter');
       metrics = require('./metrics.js');
-      KiteExpandAtCursorHighlight = require('./elements/kite-expand-at-cursor-highlight');
 
       if (!atom.inSpecMode()) {
         StateController.client = new BrowserClient('127.0.0.1', 46624, '', false);
@@ -565,8 +563,7 @@ const Kite = module.exports = {
 
     const position = editor.getLastCursor().getBufferPosition();
 
-    this.lastExpandAtCursorHighlight = new KiteExpandAtCursorHighlight();
-    this.lastExpandAtCursorHighlight.activate(editor, position);
+    this.highlightWordAtPosition(editor, position);
 
     if (!editor.getPath()) { return; }
 
@@ -586,8 +583,24 @@ const Kite = module.exports = {
     }
   },
 
-  getLastExpandAtCursorHighlight() {
-    return this.lastExpandAtCursorHighlight;
+  highlightWordAtPosition(editor, position, cls = '') {
+    if (!VirtualCursor) { VirtualCursor = require('./virtual-cursor'); }
+
+    const cursor = new VirtualCursor(editor, position);
+    const range = cursor.getCurrentWordBufferRange({
+      includeNonWordCharacters: false,
+    });
+    const marker = editor.markBufferRange(range, {
+      invalidate: 'touch',
+    });
+    const decoration = editor.decorateMarker(marker, {
+      type: 'highlight',
+      class: `expand-at-cursor-highlight ${cls}`,
+      item: this,
+    });
+
+    // Timed for all transition to be finished by then
+    setTimeout(() => decoration.destroy(), 800);
   },
 
   isSidebarVisible() {
